+++ test/dummy/app/models/user.rb
@@ -2,4 +2,5 @@
 
 class User < ApplicationRecord
   has_many :posts
+  has_and_belongs_to_many :tags
 end