+++ test/dummy/test/factories/offers_settings.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :offers_settings do
+    user
+  end
+end